% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{Svensk_antal}
\alias{Svensk_antal}
\title{Svensk antal}
\usage{
Svensk_antal(x, n = 0)
}
\arguments{
\item{x}{tal som ska fa "ratt" formatering - " " som tusentalsavg., "," som decimal.}

\item{n}{antal decimaler (nsmall). default = 0}
}
\description{
Svensk antal
}
