% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{Svensk_procent}
\alias{Svensk_procent}
\title{Svensk procent}
\usage{
Svensk_procent(x, ggr100 = TRUE, n = 0)
}
\arguments{
\item{x}{tal som ska anges som procent}

\item{ggr100}{boolean; TRUE: x ar i decimalform (och multipliceras med 100). Standard.
FALSE: x ar inte i decimalform (multpliceras inte med 100).}

\item{n}{antal decimaler (nsmall). default = 0}
}
\description{
Svensk procent
}
