% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{uka_farg}
\alias{uka_farg}
\title{Funktion som extraherar profilfarger}
\usage{
uka_farg(...)
}
\arguments{
\item{...}{char-vektor som hamtar specifika farger}
}
\description{
Funktion som ger tillgang till uka:s profilfarger, som de ar lagrade i
vektorn uka_farg_vect.
}
