% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{uka_paletter}
\alias{uka_paletter}
\title{Funktion som skapar paletter baserade pa fargerna i vektorn}
\format{
An object of class \code{list} of length 7.
}
\usage{
uka_paletter
}
\value{
uka_prim:         UKA:s primara farger (9 st)

uka_sek:          UKA:s sekundara farg (9 st)

uka_gul/bla/lila: UKA:s gula/bla/lila farger (6 var)

uka_1:            Forslag pa standardtema (12 st)
}
\description{
Funktion som skapar paletter baserade pa fargerna i vektorn
}
\keyword{datasets}
